requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.129.0
greenlet==3.3.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
xlrd==2.0.2
